Raise error when including Dry::Monads without loading all monads first
