+++ lib/dry/monads.rb
@@ -3,5 +3,12 @@ module Dry
   #
   # @api public
   module Monads
+    def self.included(base)
+      if const_defined?(:CONSTRUCTORS)
+        base.include(*CONSTRUCTORS)
+      else
+        raise "Load all monads first with require 'dry/monads/all'"
+      end
+    end
   end
 end
